refactored to take double numbers as input

+++ src/Calculator.java
@@ -1,23 +1,23 @@
 public class Calculator {
 	
-	public int add(int num1, int num2) {
+	public double add(double num1, double num2) {
 		return num1 + num2;
 	}
 	
-	public int subtract(int num1, int num2) {
+	public double subtract(double num1, double num2) {
 		return num1 - num2;
 	}
 	
-	public int multiply(int num1, int num2) {
+	public double multiply(double num1, double num2) {
 		return num1 * num2;
 	}
 	
-	public int divide(int num1, int num2) {
+	public double divide(double num1, double num2) {
 		if (num2 == 0) throw new IllegalArgumentException("Cannot divide a number by 0");
 		else return num1 / num2;
 	}
 	
-	public int square(int num1) {
+	public double square(double num1) {
 		return num1 * num1;
 	}
